Implement ah_base changes on Windows (not passing all tests)

+++ libraries/ah_http/src/http-client.c
@@ -83,7 +83,7 @@ static ah_http_client_t* s_upcast_to_client(ah_tcp_conn_t* conn)
     ah_assert_if_debug(conn != NULL);
 
     // This is only safe if `conn` is a member of an ah_http_client_t value.
-    ah_http_client_t* srv = (ah_http_client_t*) &((uint8_t*) conn)[-offsetof(ah_http_client_t, _conn)];
+    ah_http_client_t* srv = (ah_http_client_t*) &((uint8_t*) conn)[-((ptrdiff_t) offsetof(ah_http_client_t, _conn))];
 
     ah_assert_if_debug(srv->_vtab != NULL);
     ah_assert_if_debug(srv->_trans._vtab != NULL);
